Make default boot state user configurable (#30)
This pull request introduces a new `BootStatePolicy` trait to the
bootloader library, allowing bootloader applications to do the
following:
1. Set a default boot state specific to their application
2. Perform application specific validation on the current boot state. A
valid boot state from the perspective of the slimloader library may not
be valid for the application. For example, up to 7 slots are possible
but a specific application may have only 2.

The trait is defined for the `Config` structure and not the `Board`
structure since this policies are application specific and not platform
dependent.

---------

Co-authored-by: Ayodeji Ige <ayodeji.ige@microsoft.com>